Create an anagram service
Due to memory constraint, I decided to process the file in chunks.
If words are anagrams they must be same length. Because of that chunks are generated based on word length.

To collect the anagrams, I decided to generate a common key for anagrams and collect them in a Map.
I used sorted chars of the words which is same for the anagrams.

Sorting the chars' a quick solution, but it's costly due to ordering.

I would create a different key with char and count of it. (e.g. aab -> aba -> a2b1)
I could use an integer array to generate that key which ascii codes and time complexity would be O(n)

Uppercase and lowercase were were treated as two different words. "ABC" and "abc" are anagram but 2 different word.